% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calheatmap.R
\name{ggcalheatmap}
\alias{ggcalheatmap}
\title{Calendar Heatmap}
\usage{
ggcalheatmap(df, date, n, tooltip = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{date, n}{Unquoted column names.}

\item{tooltip}{Either a character string or a tooltip function. If NULL, a `ggplot2` object object is returned,
otherwise, a `ggiraph` object is returned.}
}
\description{
Plot a calendar heatmap from a data frame.
}
